fix: remove --json and root URL from bottle build
The JSON was only ever intended to get the bottle information from to put in the formula. That mainly consists of the sha256 which could easily and quickly be generated when changing that information so just remove it. That is one less file to put in S3. Same basically goes for the URL.